WS2 Grade3 super user added and documentation added, some minor fixes and smoke testing of a basic application done, A better example added

+++ Workshop2/Grade3/src/model/FileStorage.java
@@ -120,7 +120,7 @@ public class FileStorage {
 					
 							String splitter[] = data.split(";");
 							int id = Integer.parseInt (splitter[0]);
-							int persNo= Integer.parseInt (splitter[2]);
+							long persNo= Long.parseLong (splitter[2]);
 						
 							rebuildCache.initMemeber(id, splitter[1], persNo);
